+++ packages/hoast/src/utils/call.js
@@ -1,7 +1,12 @@
 // Import internal modules.
 import iterate from './iterate.js'
 
-const call = async function (options, context, functionName, ...functionArguments) {
+const call = async function (
+  options,
+  context,
+  functionName,
+  ...functionArguments
+) {
   if (Array.isArray(context)) {
     if (context.length === 0) {
       return
@@ -11,7 +16,10 @@ const call = async function (options, context, functionName, ...functionArgument
       exhausted: false,
       next: async function (index) {
         const contextItem = context[index]
-        if (typeof (contextItem) === 'object' && typeof (contextItem[functionName]) === 'function') {
+        if (
+          typeof (contextItem) === 'object' &&
+          typeof (contextItem[functionName]) === 'function'
+        ) {
           await contextItem[functionName](...functionArguments)
         }
 
@@ -23,7 +31,10 @@ const call = async function (options, context, functionName, ...functionArgument
     return
   }
 
-  if (typeof (context) === 'object' && typeof (context[functionName]) === 'function') {
+  if (
+    typeof (context) === 'object' &&
+    typeof (context[functionName]) === 'function'
+  ) {
     await context[functionName](...functionArguments)
   }
 }