+++ packages/utils/importVersion.js
@@ -1,7 +1,20 @@
-export default async (filePath, version = 0) => {
-  // Append version to file path.
-  if (version && version > 0) {
+export default async (
+  filePath,
+  version = 0,
+  namespace = null,
+) => {
+  if (
+    version && (
+      typeof (version) === 'number'
+        ? version > 0
+        : true
+    )
+  ) {
     filePath += '?version=' + version
+  }
+
+  if (namespace) {
+    filePath += '?namespace=' + namespace
   }
 
   return await import(filePath)