+++ packages/base-source/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+## 3.1.0
+
+- Added a `fence` option to split an item up into separate items.