+++ packages/process-handlebars/CHANGELOG.md
@@ -1,1 +1,6 @@
 # Changelog
+
+##
+
+- Append `.hbs` suffix to template paths if not found with at first.
+- Remove `.js` suffix from helpers.