+++ packages/source-readfiles/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+## 3.2.0
+
+- Upgraded `BaseSource` dependency.